Rename icf to source in writer
